# api/openapi/v1.yaml
components:
  schemas:
...
    github_com_denisakp_ogoune_internal_dto_v1.AddTagsRequest:
      properties:
        tag_ids:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    github_com_denisakp_ogoune_internal_dto_v1.CreateMonitorRequest:
      properties:
        component_id:
          type: string
        interval:
          type: integer
        keyword:
          type: string
        name:
          type: string
        protocol_port:
          type: integer
        protocol_type:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        target:
          type: string
        timeout:
          type: integer
        type:
          type: string
      type: object
...
    github_com_denisakp_ogoune_internal_dto_v1.ErrorDetail:
      properties:
        code:
          type: string
        fields:
          items:
            $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.FieldError'
          type: array
          uniqueItems: false
        message:
          type: string
      type: object
    github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorDetail'
      type: object
    github_com_denisakp_ogoune_internal_dto_v1.FieldError:
      properties:
        field:
          type: string
        message:
          type: string
      type: object
...
    github_com_denisakp_ogoune_internal_dto_v1.MetaResponse:
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
      type: object
    github_com_denisakp_ogoune_internal_dto_v1.MonitorResponse:
      properties:
        component_id:
          type: string
        created_at:
          type: string
        host_id:
          description: 'spec 079: optional link to a monitored host (null when unlinked)'
          type: string
        id:
          type: string
        interval:
          type: integer
        last_checked_at: {}
        name:
          type: string
        status:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        target:
          type: string
        timeout:
          type: integer
        type:
          type: string
        updated_at:
          type: string
      type: object
    github_com_denisakp_ogoune_internal_dto_v1.MonitorUptimeStat:
      properties:
        hour:
          type: string
        successful_count:
          type: integer
        total_count:
          type: integer
        uptime_percent:
          type: number
      type: object
    github_com_denisakp_ogoune_internal_dto_v1.MonitorUptimeStatsResponse:
      properties:
        resource_id:
          type: string
        stats:
          items:
            $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.MonitorUptimeStat'
          type: array
          uniqueItems: false
      type: object
...
    github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorResponse:
      properties:
        data:
          $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.MonitorResponse'
        meta:
          $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.MetaResponse'
      type: object
    github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorUptimeStatsResponse:
      properties:
        data:
          $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.MonitorUptimeStatsResponse'
        meta:
          $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.MetaResponse'
      type: object
...
    github_com_denisakp_ogoune_internal_dto_v1.UpdateMonitorRequest:
      properties:
        component_id:
          type: string
        interval:
          type: integer
        keyword:
          type: string
        name:
          type: string
        protocol_port:
          type: integer
        protocol_type:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        target:
          type: string
        timeout:
          type: integer
        type:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: Ogoune uptime monitoring — Public REST API v1.
  title: Ogoune Public API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /monitors:
    get:
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        schema:
          type: integer
      - description: Items per page (1-100, default 20)
        in: query
        name: per_page
        schema:
          type: integer
      - description: Filter by exact tag name
        in: query
        name: tag
        schema:
          type: string
      - description: Filter by monitor type (http|tcp|dns|icmp|keyword|protocol|heartbeat)
        in: query
        name: type
        schema:
          type: string
      - description: Filter by active state (default true)
        in: query
        name: is_active
        schema:
          type: boolean
      - description: Substring search on name + target (case-insensitive)
        in: query
        name: q
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Unauthorized
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Unprocessable Entity
      security:
      - BearerAuth: []
      summary: List monitors
      tags:
      - monitors
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.CreateMonitorRequest'
                description: Monitor payload
                summary: body
        description: Monitor payload
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorResponse'
          description: Created
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Unprocessable Entity
      security:
      - BearerAuth: []
      summary: Create a monitor
      tags:
      - monitors
  /monitors/{id}:
    delete:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete a monitor
      tags:
      - monitors
    get:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get a monitor by ID
      tags:
      - monitors
    patch:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.UpdateMonitorRequest'
                description: Partial update payload
                summary: body
        description: Partial update payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Unprocessable Entity
      security:
      - BearerAuth: []
      summary: Partially update a monitor
      tags:
      - monitors
    put:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.UpdateMonitorRequest'
                description: Update payload
                summary: body
        description: Update payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update a monitor
      tags:
      - monitors
...
  /monitors/{id}/pause:
    post:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Pause a monitor
      tags:
      - monitors
  /monitors/{id}/resume:
    post:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Resume a monitor
      tags:
      - monitors
  /monitors/{id}/tags:
    post:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.AddTagsRequest'
                description: Tag IDs
                summary: body
        description: Tag IDs
        required: true
      responses:
        "204":
          description: No Content
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Attach tags to a monitor
      tags:
      - monitors
  /monitors/{id}/tags/{tagID}:
    delete:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Tag ID
        in: path
        name: tagID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Detach a tag from a monitor
      tags:
      - monitors
  /monitors/{id}/uptime-stats:
    get:
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.SingleResponse-github_com_denisakp_ogoune_internal_dto_v1_MonitorUptimeStatsResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_denisakp_ogoune_internal_dto_v1.ErrorResponse'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Hourly uptime statistics for a monitor
      tags:
      - monitors
...
servers:
- url: localhost:9596/api/v1